+++ node/clients/search.ts
@@ -1,47 +1,63 @@
+import { ExternalClient, InstanceOptions, IOContext } from '@vtex/api'
 
-const axios = require("axios");
-
-import {
-  ExternalClient,
-  InstanceOptions,
-  IOContext
-} from '@vtex/api'
+const axios = require('axios')
 
 interface RefIdArgs {
   refids: any
 }
 
 export class Search extends ExternalClient {
-
   public constructor(ctx: IOContext, opts?: InstanceOptions) {
-    super(`http://${ctx.account}.vtexcommercestable.com.br/`, ctx, opts);
-  }  
-    
-  public skuFromRefIds = async ({refids}:RefIdArgs) => {
-  
-    const url = `http://${this.context.account}.vtexcommercestable.com.br/api/catalog_system/pub/sku/stockkeepingunitidsbyrefids`;
+    super(`http://${ctx.account}.vtexcommercestable.com.br/`, ctx, opts)
+  }
 
-    const res = await axios.default.post(url, refids,{headers: {
-      "Content-Type": "application/json",
-      "Authorization": `bearer ${this.context.authToken}`
-    }});
+  public skuFromRefIds = async ({ refids }: RefIdArgs) => {
+    const url = `http://${this.context.account}.vtexcommercestable.com.br/api/catalog_system/pub/sku/stockkeepingunitidsbyrefids`
 
-    let result:any = [];
+    const res = await axios.default.post(url, refids, {
+      headers: {
+        'Content-Type': 'application/json',
+        Authorization: `bearer ${this.context.authToken}`,
+      },
+    })
 
-    if(res.status === 200) {
-      
-      const refs = Object.getOwnPropertyNames(res.data);
+    const result: any = []
+
+    if (res.status === 200) {
+      const refs = Object.getOwnPropertyNames(res.data)
 
       refs.forEach(id => {
         result.push({
           sku: res.data[id],
-          refid: id
+          refid: id,
         })
-      });
-    
+      })
     }
 
-    return result;
+    return result
   }
-  
+
+  public sellers = async () => {
+    const url = `http://${this.context.account}.vtexcommercestable.com.br/api/catalog_system/pvt/seller/list`
+
+    const res = await axios.default.get(url, {
+      headers: {
+        'Content-Type': 'application/json',
+        Authorization: `bearer ${this.context.authToken}`,
+      },
+    })
+
+    let result: any = []
+
+    if (res.status === 200) {
+      result = res.data.map(({ SellerId, Name }: any) => {
+        return {
+          id: SellerId,
+          name: Name,
+        }
+      })
+    }
+
+    return result
+  }
 }